added functionality for all component types
