+++ source/GameInterface/Services/PartyComponents/Messages/PartyComponentCreated.cs
@@ -1,13 +1,8 @@
 ﻿using Common.Messaging;
-using GameInterface.Services.PartyComponents.Data;
 using TaleWorlds.CampaignSystem.Party.PartyComponents;
 
 namespace GameInterface.Services.PartyComponents.Messages;
-internal class PartyComponentCreated : IEvent
+internal record PartyComponentCreated(PartyComponent Instance) : IEvent
 {
-    public PartyComponent Instance { get; }
-    public PartyComponentCreated(PartyComponent instance)
-    {
-        Instance = instance;
-    }
+    public PartyComponent Instance { get; } = Instance;
 }
